(PC-37097)[PRO] refactor: move DialogStockThingDeleteConfirm up
