+++ pro/src/pages/IndividualOffer/components/DialogStockThingDeleteConfirm.tsx
